Fixing tests and API.
