+++ amplify/backend/function/api/test.sh
@@ -1,4 +1,4 @@
 cd "$(dirname $0)"
 cd src
 tsc --project tsconfig.json
-TZ='UTC' npx mocha -b --timeout 26s --exit --trace-warnings test/**/*.test.js
+STORAGE_JSONSTORAGE_BUCKETNAME=json183906-dev TZ='UTC' npx mocha -b --timeout 180s --exit --trace-warnings test/**/*.test.js
